+++ src/qs_codec/__init__.py
@@ -1,9 +1,9 @@
 """A query string encoding and decoding library for Python. Ported from qs_codec for JavaScript."""
 
-__version__ = "1.1.4"
+__version__ = "1.1.5"
 
-from .decode import decode
-from .encode import encode
+from .decode import decode, load, loads
+from .encode import dumps, encode
 from .enums.charset import Charset
 from .enums.duplicates import Duplicates
 from .enums.format import Format
@@ -17,6 +17,9 @@ from .models.undefined import Undefined
 __all__ = [
     "decode",
     "encode",
+    "dumps",
+    "loads",
+    "load",
     "Charset",
     "Duplicates",
     "Format",